requirements.txt
Django
djangorestframework
drf-spectacular
psycopg2-binary
gunicorn
python-dotenv
dotenv
uwsgi
flake8
numpy
meson